+++ src/scenarioRunner/evaluateScenario.ts
@@ -7,9 +7,17 @@ import type { Screenshot } from "../types/Screenshot";
 export async function evaluateScenarioByPuppeteer(
 	page: puppeteer.Page,
 	playlogJsonPath: string,
+	gameJsonPath?: string,
 	takeScreenshotFunc?: (s: Screenshot) => void,
 	selector: string = "canvas"
 ): Promise<void> {
+	if (gameJsonPath) {
+		// game.jsonの動的読み込みのため、require の lint エラーを抑止
+		/* eslint-disable @typescript-eslint/no-require-imports */
+		const gameJson = require(gameJsonPath);
+		// コンテンツ描画領域が他の要素と重ならないようにするために、画面サイズをコンテンツのサイズよりも大きくする
+		await page.setViewport({ width: Math.round(1.1 * gameJson.width), height: Math.round(1.1 * gameJson.height) });
+	}
 	await page.waitForSelector(selector);
 	// canvas 要素が存在する環境で実行する前提なので、! を付与する
 	const canvasHandle = (await page.$(selector))!;