+++ src/scenarioRunner/serve/createServeScenarioRunner.ts
@@ -17,8 +17,13 @@ import { createAkashicServe } from "./AkashicServe";
 // コンテンツ実行時間の上限。テスト時に実行時間がこの長さを超える場合、確実に失敗することに注意が必要
 const CONTENT_LIMIT_MAX_TIME = 300000;
 
-export async function createServeScenarioRunner(binSrc: TargetBinarySource): Promise<ScenarioRunner> {
-	const serveBin = await createAkashicServe(binSrc);
+interface CreateServeScenarioRunnerParameterObject {
+	type: "serve" | "export-zip";
+	binSrc: TargetBinarySource;
+}
+
+export async function createServeScenarioRunner(param: CreateServeScenarioRunnerParameterObject): Promise<ScenarioRunner> {
+	const serveBin = await createAkashicServe(param.binSrc);
 	return {
 		run: async (
 			contentDirPath: string,
@@ -105,13 +110,21 @@ export async function createServeScenarioRunner(binSrc: TargetBinarySource): Pro
 					}
 				}
 			} catch (e) {
+				// TODO: この辺りのエラーハンドリングは他のScenarioRunnerとほぼ同じコードになっているので、「シナリオを実行してエラー時にスクリーンショットを撮る」一連の流れを共通化すべき
 				if (e instanceof TimeoutError) {
-					const timeoutImage: Screenshot = {
-						fileName: `timeout_try${playCount}_${extractDirname(scenarioPath)}.png`,
+					const screenshot: Screenshot = {
+						fileName: `${param.type}_timeout_try${playCount}_${extractDirname(scenarioPath)}.png`,
 						base64: await page.screenshot({ encoding: "base64" })
 					};
-					return { status: "timeout", timeoutImage };
+					return { status: "timeout", screenshot };
 				} else {
+					if (page && !page.isClosed()) {
+						const screenshot: Screenshot = {
+							fileName: `${param.type}_error_try${playCount}_${extractDirname(scenarioPath)}.png`,
+							base64: await page.screenshot({ encoding: "base64" })
+						};
+						return { status: "error", screenshot, error: e };
+					}
 					throw e;
 				}
 			} finally {